+++ src/app.tsx
@@ -1,9 +1,9 @@
 import React from 'react'
-import Footer from './components/footer/footer'
-import Header from './components/header/header'
+import Footer from './components/footer/footer.jsx'
+import Header from './components/header/header.jsx'
 import './app.css'
-import { useHelia } from './hooks/use-helia'
-import { Page } from './page/page'
+import { useHelia } from './hooks/use-helia.js'
+import { Page } from './page/page.jsx'
 
 export const App = (): React.JSX.Element => {
   const { starting, error } = useHelia()
